requirements.txt
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.5
google-api-core==2.30.0
google-auth==2.49.1
google-cloud-bigquery==3.40.1
google-cloud-core==2.5.0
google-cloud-storage==3.10.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.78.0
idna==3.11
packaging==26.0
proto-plus==1.27.1
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.32.5
six==1.17.0
typing_extensions==4.15.0
urllib3==2.6.3
